recommendation

+++ application/vacancy/repository.go
@@ -12,7 +12,7 @@ type RepositoryVacancy interface {
 	GetVacancyList(uint, uint, uuid.UUID, int) ([]models.Vacancy, error)
 	SearchVacancies(models.VacancySearchParams) ([]models.Vacancy, error)
 	AddRecommendation(uuid.UUID, int) error
-	GetRecommendation(start int, limit int, salary *int, spheres []int) ([]models.Vacancy, error)
+	GetRecommendation(start int, limit int, salary *float64, spheres []int) ([]models.Vacancy, error)
 	GetPreferredSpheres(userID uuid.UUID) ([]Pair, error)
-	GetPreferredSalary(uuid.UUID) (*int, error)
+	GetPreferredSalary(uuid.UUID) (*float64, error)
 }
